Module 10 - Adding Error Handling and Validation

+++ Application/Activities/Commands/DeleteActivity.cs
@@ -1,24 +1,34 @@
 ﻿using MediatR;
 using Persistence;
 using Domain;
+using Application.Core;
 
 namespace Application.Activities.Commands
 {
     public class DeleteActivity
     {
-        public class Command : IRequest
+        public class Command : IRequest<Result<Unit>> // Unnit is used to represent a void return type in MediatR
         {
             public required string Id { get; set; }
         }
 
-        public class Handler(AppDbContext context) : IRequestHandler<Command>
+        public class Handler(AppDbContext context) : IRequestHandler<Command, Result<Unit>>
         {
-            public async Task Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
-                var activity = await context.Activities.FindAsync([request.Id], cancellationToken) ?? throw new Exception("Activity not found");
-                context.Activities.Remove(activity);
-                await context.SaveChangesAsync(cancellationToken);
+                var activity = await context.Activities.FindAsync([request.Id], cancellationToken);
 
+                if (activity == null)
+                {
+                    return Result<Unit>.Failure("Activity not found", 404);
+                } 
+
+                context.Activities.Remove(activity);
+                var result = await context.SaveChangesAsync(cancellationToken) > 0;
+
+                if (!result) return Result<Unit>.Failure("Failed to delete activity", 400);
+
+                return Result<Unit>.Success(Unit.Value);
             }
         }
     }
